+++ plom/client/annotator.py
@@ -367,7 +367,6 @@ class Annotator(QWidget):
             self.modeInformation.append((key, tab))
 
         # after grabbed mode information, reset rubric_widget
-        self.rubric_widget.reset()
         self.rubric_widget.setEnabled(False)
 
         del self.scene
@@ -386,6 +385,8 @@ class Annotator(QWidget):
         self,
         tgvID,
         question_label,
+        version,
+        max_version,
         testName,
         paperdir,
         saveName,
@@ -399,9 +400,12 @@ class Annotator(QWidget):
         Args:
             tgvID (str): Test-Group-Version ID code.  For example, for
                 Test #0027, group #13, version #2, we have `t0027g13v2`.
+                TODO: currently only `t0027g13`, no version, despite name.
             question_label (str): The name of the question we are
                 marking.  This is generally used for display only as
                 there is an integer for precise usage.
+            version (int): which version are we working on?
+            max_version (int): what is the largest version in this assessment?
             testName (str): Test Name
             paperdir (dir): Working directory for the current task
             saveName (str/pathlib.Path): file name (and dir, optionally)
@@ -421,6 +425,8 @@ class Annotator(QWidget):
         """
         self.tgvID = tgvID
         self.question_num = int(re.split(r"\D+", tgvID)[-1])
+        self.version = version
+        self.max_version = max_version
         self.question_label = question_label
         self.testName = testName
         s = "{} of {}: {}".format(self.question_label, testName, tgvID)
@@ -449,7 +455,8 @@ class Annotator(QWidget):
         self.rubric_widget.changeMark(
             self.getScore(), self.getMarkingState(), self.maxMark
         )
-        self.rubric_widget.setQuestionNumber(self.question_num)
+        self.rubric_widget.setQuestion(self.question_num, self.question_label)
+        self.rubric_widget.setVersion(self.version, self.max_version)
         self.rubric_widget.setEnabled(True)
 
         # TODO: Make handling of rubric less hack.