+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * "Tech preview" of a web-interface for configuring a Plom server.
 * Support for external databases.
+* Rubrics can be version-specific.
 
 ### Changed
 * Server stores annotation data in the database instead of in `.plom` files.